Move isOutMostRow() implementation to sub-classes
